chore(events_trace): Create TraceTestCase case
This takes most of the logic from OrganizationEventsTraceEndpointBase to testutils.
